requirements.txt
asgiref==3.7.2
Django==3.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.1.1
PyJWT==2.8.0
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.7.1
